Completely extract GameInteractionHandler into multiple classes
